export_requirements.txt
apscheduler==3.10.4
boto3==1.34.96
botocore==1.34.96
email-validator==2.1.0.post1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==23.0.0
psutil==5.9.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
trafilatura==1.6.3
Werkzeug==3.0.2
WTForms==3.1.1